+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 - `test_data_tools` has tests for almost all classes and functions.
   Functions that include plots are skipped (cannot be tested). Only test
   missing is `models.Lasso`.
+- In `diffusion` module:
+    - Function `euler_explicit2D`: Computes diffusion on a 2D space over
+      a time-step using Euler explicit method.
+    - Function `euler_implicit_coef_mat`: Computes the coefficient
+      matrix to solve the diffusion problem with the Euler implicit
+      method.
+    - Function `crank_nicolson_coef_mats`: Computes the coefficient
+      matrices to solve the diffusion problem with the Crank-Nicolson
+      method.
+    - Function `build_coef_mat`: Builds a coefficient matrix according
+      to the central and neighbor coefficients, system size and boundary
+      conditions.
 - In `models` module:
     - Class `DoseResponse` improved and bugs fixed.
 - In `databases` module: